A day's work
Here is all the changes off the top of my head:

 - Lots of atomics have been changed from the default ceq_con to either
acquire, release or relaxed (depending on which is necessary)

 - We shouldn't do jobs bug has been fixed by simply waiting till we can
do it. It's a timing bug which I stopped being able to reproduce after the next
fix, now it's been fixed by simply waiting for it to be ready.

 - ConditionalContinue had a bug where signal could be incremented after
the check in wait but notify_all could be called before the other thread
reached wait. Moving the mutex back to also protect the signal check
fixes this.

 - Some mutexes have been replaced with atomics where appropriate.
 Missing mutexes have been added as I've noticed them.

 - WorkerRoutine has been simplified.

 - When pulling a dep job, instead of swapping it with the last element
 then popping it back we now swap it with nullptr. If we notice a
 nullptr at the end we pop that. Make runnable has been changed to
 simply submit a job pulled by the previously mentioned function.

Signed-off-by: Hal Gentz <zegentzy@protonmail.com>